requirements.txt
azure-identity
requests
PyYAML
pydantic
